backend/requirements.txt
faiss-cpu
Flask
flask-cors
fsspec
numpy
PyGithub
requests
langchain
google-generativeai 